delphi/requirements.txt
langgraph==1.1.6
langchain-core==1.3.0
langsmith==0.7.32
google-genai==1.72.0
streamlit
supabase==2.7.4
pydantic==2.12.4
python-dotenv==1.2.1
httpcore==1.0.9
httpx==0.28.1
pytest==9.0.2
ruff
